I've updated the package_sire framework so that it now works correctly with the miniconda install. This needs some testing but it looks ok from what I have run so far...

+++ wrapper/build/copy_into_conda.py
@@ -53,6 +53,9 @@ for file in os.listdir(fromdir):
                (sys.executable,sys.argv[0],sys.argv[1],file,sys.argv[2],file))
 
     if os.path.exists(tofile):
+        if os.path.isdir(tofile):
+            continue
+
         if force:
             os.remove(tofile)
         else:
